feat: Improve GitHub Copilot workshop instructions and titles
Update the workshop instructions to include a step-by-step guide for downloading the workshop repository using either Git command or by downloading the ZIP file. Additionally, update the workshop title to "GitHub Copilot Workshop" and the short title to "GH Copilot Workshop". Also, add the tags "GitHub Copilot" and "python" in addition to "javascript" for better filtering and searching.

Co-authored-by: Kaisei Tsujimoto <katsujim@microsoft.com>